% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sc_cpp_recresid}
\alias{sc_cpp_recresid}
\title{Computation of recursive residuals in C++}
\usage{
sc_cpp_recresid(X, y, start, end, tol)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{start}{integer (1-based) index of the first observation to compute recursive residuals}

\item{end}{integer (1-based) index of the last observation to compute recursive residuals}

\item{tol}{tolerance in the computation of recursive model coefficients}
}
\value{
vector containing the recursive residuals
}
\description{
Computation of recursive residuals in C++
}
\seealso{
\code{\link{recresid}} and \code{\link{recresid.default}}
}

